Adding a file of alternate T-P parametrizations
The new "user" directory contains the default values for Apollo.py, which are now just imported en masse.

It also contains a file of functions which can be plugged in to the "layers" pipeline for T-P profiles. The idea is you just pass the list of "atm" arguments in whatever form. The appropriate function spits back an array of temperatures.

Also, changed some of the hard-coded values for constants to variables defined near the top of Apollo.py